refactor: remove api.rs from massa-models
