+++ massa-api/src/private.rs
@@ -25,9 +25,9 @@ use massa_models::output_event::SCOutputEvent;
 use massa_models::prehash::PreHashSet;
 use massa_models::{
     address::Address,
-    api::EventFilter,
     block::{Block, BlockId},
     endorsement::EndorsementId,
+    execution::EventFilter,
     operation::OperationId,
     slot::Slot,
 };
