+++ massa-execution-exports/src/controller_traits.rs
@@ -7,8 +7,8 @@ use crate::ExecutionError;
 use crate::{ExecutionAddressInfo, ReadOnlyExecutionOutput};
 use massa_models::address::Address;
 use massa_models::amount::Amount;
-use massa_models::api::EventFilter;
 use massa_models::block::BlockId;
+use massa_models::execution::EventFilter;
 use massa_models::operation::OperationId;
 use massa_models::output_event::SCOutputEvent;
 use massa_models::prehash::PreHashMap;
