+++ massa-models/src/execution.rs
@@ -3,27 +3,6 @@
 use crate::{address::Address, operation::OperationId, slot::Slot};
 use serde::{Deserialize, Serialize};
 
-/// Block status within the graph
-#[derive(Eq, PartialEq, Debug, Deserialize, Serialize)]
-pub enum BlockGraphStatus {
-    /// received but not yet graph-processed
-    Incoming,
-    /// waiting for its slot
-    WaitingForSlot,
-    /// waiting for a missing dependency
-    WaitingForDependencies,
-    /// active in alternative cliques
-    ActiveInAlternativeCliques,
-    /// active in blockclique
-    ActiveInBlockclique,
-    /// forever applies
-    Final,
-    /// discarded for any reason
-    Discarded,
-    /// not found in graph
-    NotFound,
-}
-
 /// filter used when retrieving SC output events
 #[derive(Default, Debug, Deserialize, Clone, Serialize)]
 pub struct EventFilter {
@@ -49,19 +28,4 @@ pub struct EventFilter {
     /// Some(false) means events coming from a succeeded sc execution
     /// None means both
     pub is_error: Option<bool>,
-}
-
-/// When an address is drawn to create an endorsement it is selected for a specific index
-#[derive(Debug, Deserialize, Serialize, Hash, PartialEq, Eq)]
-pub struct IndexedSlot {
-    /// slot
-    pub slot: Slot,
-    /// endorsement index in the slot
-    pub index: usize,
-}
-
-impl std::fmt::Display for IndexedSlot {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        writeln!(f, "Slot: {}, Index: {}", self.slot, self.index)
-    }
 }
