+++ massa-pos-exports/src/controller_traits.rs
@@ -7,9 +7,10 @@ use std::collections::BTreeMap;
 
 use crate::PosResult;
 use massa_hash::Hash;
-use massa_models::address::Address;
-use massa_models::api::IndexedSlot;
-use massa_models::slot::Slot;
+use massa_models::{
+    address::Address,
+    slot::{IndexedSlot, Slot},
+};
 
 #[cfg(feature = "testing")]
 use std::collections::{HashMap, VecDeque};
