+++ massa-pos-exports/src/test_exports/mock.rs
@@ -10,7 +10,7 @@ use std::{
 };
 
 use massa_hash::Hash;
-use massa_models::{address::Address, api::IndexedSlot, slot::Slot};
+use massa_models::{address::Address, execution::IndexedSlot, slot::Slot};
 
 use crate::{PosResult, Selection, SelectorController};
